Move Markov chain macro atom to separate file
